requirements.txt
dotenv>=0.9.9
fastmcp>=2.10.6
google-adk>=1.7.0
scikit-learn==1.6.1
langchain-community==0.3.20
wikipedia==1.4.0
crewai-tools==0.38.1
dateparser==1.2.1
mcp==1.10.1
lxml
beautifulsoup4
fire==0.7.0
google-cloud-aiplatform[adk,agent_engines]==1.88.0
langchain-community==0.3.20
wikipedia==1.4.0
